Rename controllers to controller
