+++ internal/controller/imageupdateautomation_controller.go
