+++ internal/controller/imageupdateautomation_controller_test.go
