+++ internal/controller/testdata/pathconfig/no/deploy.yaml
