+++ internal/controller/testdata/pathconfig/yes/deploy.yaml
